Redo validations
Redo validations, now only do simple change from parsing long instead of
int
Also cleanup wrongly changed usings

Conflicts:
	src/Hangfire.Core/Dashboard/JobHistoryRenderer.cs

+++ src/Hangfire.Core/Dashboard/JobHistoryRenderer.cs
@@ -129,7 +129,8 @@ namespace Hangfire.Dashboard
 
             if (stateData.ContainsKey("Latency"))
             {
-                var latency = TimeSpan.FromMilliseconds(int.Parse(stateData["Latency"]));
+                var latency = TimeSpan.FromMilliseconds(long.Parse(stateData["Latency"]));
+
                 builder.AppendFormat("<dt>Latency:</dt><dd>{0}</dd>", html.ToHumanDuration(latency, false));
 
                 itemsAdded = true;
